### 2020-06-30 #2380 @galessiorob
👋 Hi @LukasAdamowicz!

I am in the process of assigning reviewers to this submission, but if you have any suggestions for reviewers please let me know in a comment without tagging them directly.

In the meantime, it looks like you're missing some DOIs, if you want to start adding those that would be great.

Many thanks!
______________

**Basic checks:**

* [x]  There’s a general description of the software
* [x] The software is within scope as research software
* [x] OSI-approved license: MIT License
