### 2020-06-30 #2380 @danielskatz
@galessiorob  - when you agree to edit a submission, you should
(or @whedon assign me as editor)
and if the submission has been waitlisted, remove the waitlist label
I'll do this now for your 2 that you agreed to edit
