### 2020-07-08 #2380 @jsgalan
Hi all,

I made an initial install, all the requirements were satisfied

Two test failed as can be seen:

and one .h5 file was missing somehow.

should this example data should be downloaded from anywhere else?

Best,
