### 2020-07-08 #2449 @whedon
[REVIEW]: Sit2StandPy: An Open-Source Python Package for Detecting and Quantifying Sit-to-Stand Transitions Using an Accelerometer on the Lower Back'
**Submitting author:** @LukasAdamowicz (<a href="http://orcid.org/0000-0001-6992-9545">Lukas Adamowicz</a>)
**Repository:** <a href="https://github.com/PfizerRD/sit2standpy" target ="_blank">https://github.com/PfizerRD/sit2standpy</a>
**Version:** v1.0.0
