### 2020-07-08 #2449 @LukasAdamowicz
Hi all,
Thank you for the comments/work so far - DOIs were added and look like they are showing up, I'll look into making sure that the file is packed with the installation @jsgalan
