### 2020-07-08 #2449 @LukasAdamowicz
@jsgalan you should be able to now re-install from pip (make sure you get version 1.1.1.post4), which should contain the files needed for the tests (also made changes so that all the tests that were supposed to run are found correctly)
