### 2020-07-08 #2449 @gweissman
I was able to install the package easily without apparent errors. To assess functionality, I ran the accompanying tests using `pytest` and all passed. I did not import the library directly or try running any of my own code with it.

One scientific concern is that the validation of the algorithms implemented in this software package appear to have been reported in a paper that is currently under review. Therefore, the scientific validity of the algorithms can't be assessed here. If that is important for JOSS, then I would suggest pending this review until those results have passed through some peer review process.

One ethical concern is that the repo appears at the Pfizer site, and I did not see any statement of funding regarding the research or disclosure about the involvement of the funding source in the development, testing, or reporting of the software or associated manuscript.
