### 2020-07-09 #2449 @LukasAdamowicz
Thank you for your comments, @gweissman.

Regarding your last point, I added the following statement to the paper, which has been on the other papers our group has published ([GaitPy](https://joss.theoj.org/papers/10.21105/joss.01778), [SleepPy](https://joss.theoj.org/papers/10.21105/joss.01663))

> # Acknowledgements
> The Digital Medicine & Translational Imaging group at Pfizer, Inc supported the development of this package.
