### 2020-07-14 #2449 @galessiorob
Thank you so much for your review @gweissman! ⚡ I am looking into policies regarding [this comment](https://github.com/openjournals/joss-reviews/issues/2449#issuecomment-655755325), I'll update the issue with my findings.

@jsgalan checking in, I know you started the review back in the pre-review issue, feel free to copy over your comments and start checking off the list at the top assigned to you. I tried installing today and didn't run [into any issues.](https://github.com/openjournals/joss-reviews/issues/2449#issuecomment-655534519)

Let me know if I can help with anything and thanks again for your work.
