### 2020-07-14 #2449 @jsgalan
Hi all, I proceed to clone the git

Then created a Venv

Then did an initial install try (Python 2.7)

Then installed pip

Then a second install with python 3.8

Then did a manual installation of the requierements

Then installed Pytest

then ran the tests

V1

V2

All passed without problems!

Then went and did ran the code in the repository for V2 usage and old usage (here called test2 and test3, respectively)

I try in my editor, just after a fresh pip install

![Screen Shot 2020-07-14 at 11 24 18 AM](https://user-images.githubusercontent.com/971958/87450806-9e667c80-c5c4-11ea-9373-2fba8748e2ef.png)

and got the same errors:

![Screen Shot 2020-07-14 at 11 24 26 AM](https://user-images.githubusercontent.com/971958/87450809-9f97a980-c5c4-11ea-8eec-78eb472b9b82.png)
![Screen Shot 2020-07-14 at 11 24 39 AM](https://user-images.githubusercontent.com/971958/87450812-a0c8d680-c5c4-11ea-92bc-0de7b8844d4c.png)

Not sure what is going wrong of if I am missing a step somewhere.

Best
