### 2020-07-14 #2449 @LukasAdamowicz
@jsgalan thank you for finding those issues. I've addressed both of them with the usage examples, which have been updated in the repository README. In summary, there was a typo in the V1 (old usage) example, which should be fixed now:

`
...
with resources.path('sit2standpy.data', 'sample.csv') as file_path:
    ...`

Regarding V2 usage, `data_transform_function` was intended to be the users own function for achieving the required format. The comments around this have been updated, and I added a link to a gist as well, that provides a simple data loader function

[data loader GIST](https://gist.github.com/LukasAdamowicz/b8481ef32e4beeb77c80f29f34c8045e)
