### 2020-07-14 #2449 @jsgalan
Hi @LukasAdamowicz,

V1 works fine

V2 using the loader as suggested give me an error

![Screen Shot 2020-07-14 at 12 27 35 PM](https://user-images.githubusercontent.com/971958/87457241-aa0a7100-c5cd-11ea-8e83-15ced1433f28.png)

After an small inspection I realized there was an error is 'Unix Time' instead of 'Unix Timestamps' (I know it is kind off-topic since is just an example)

![Screen Shot 2020-07-14 at 12 56 46 PM](https://user-images.githubusercontent.com/971958/87459816-81847600-c5d1-11ea-828c-0cc7689723b8.png)

After that correction I get to this error, not sure why?

![Screen Shot 2020-07-14 at 12 49 57 PM](https://user-images.githubusercontent.com/971958/87459709-60238a00-c5d1-11ea-996d-a129e1bd0fa1.png)

Best
