### 2020-07-14 #2449 @LukasAdamowicz
@jsgalan thank you, apologies for these difficulties. I've gone an updated the README with the v2 usage (tested on my end) - basically what was missing was the library prefix for the `WindowDays` (and following calls) that got added to the Sequence, so it becomes

```
...
sequence.add(s2s.v2.WindowDays(hours=[8, 20]))
sequence.add(s2s.v2.AccelerationFilter())
sequence.add(s2s.v2.Detector(stillness_constraint=True))
...
```

I also fixed the gist as well to correctly use the correct key names
