### 2020-07-14 #2449 @jsgalan
HI about the written articles, I am not sure I am seeing the last version. Since I am not seeing the acknowledgment you mentioned earlier.

> Thank you for your comments, @gweissman.
>
> Regarding your last point, I added the following statement to the paper, which has been on the other papers our group has published ([GaitPy](https://joss.theoj.org/papers/10.21105/joss.01778), [SleepPy](https://joss.theoj.org/papers/10.21105/joss.01663))
>
> > # Acknowledgements
> > The Digital Medicine & Translational Imaging group at Pfizer, Inc supported the development of this package.

Also I think this can be easily solved, just adding a line to the code.

![Screen Shot 2020-07-14 at 3 13 53 PM](https://user-images.githubusercontent.com/971958/87472476-99fe8b80-c5e5-11ea-9add-2fde22e3cf4e.png)

As @gweissman has made me notice in his review, I think the authors should give a few more details on other (open souse) software implementations, or state that is a novel implementation, to be able to answer this _State of the field: Do the authors describe how this software compares to other commonly-used packages?_
