### 2020-07-17 #2449 @jsgalan
Hi Lukas, Please check the Availability Section seems to have the same error I noted earlier.

![87472476-99fe8b80-c5e5-11ea-9add-2fde22e3cf4e](https://user-images.githubusercontent.com/971958/87735219-88aab000-c79a-11ea-9c06-9083675a39f5.png)
