### 2020-07-17 #2449 @LukasAdamowicz
Fixed. I also added a statement regarding existing implementations/software of sit-to-stand algorithms
