### 2020-07-19 #2449 @jsgalan
Hi all, I just checked the last version of the article and am satisfied with the results.

I am happy to accept the submission.

Best,
