### 2020-07-29 #2449 @galessiorob
Hey @LukasAdamowicz

I talked to other editors and the consensus was that it's okay to publish the paper despite it citing an algorithm that is still not published as long as:
1. You can confirm that the other publication for which you are also an author (_Assessment of sit to stand during daily life using an accelerometer on the lower back. IEEE Journal of Biomedical and Health Informatics_) is (at least partly) focused on determining the validity/applicability of the algorithm for this area of research.
2. You can confirm this other paper is not simply about the software as well, as this might constitute self-plagiarism.
3. Finally, adding some complementary documentation around the algorithm itself:
     - State what the algorithm is in the documentation.
     - Demonstrate that their implementation is verifiable through tests.
     - Discuss the role the other manuscript plays in vetting the application of the algorithm.

Let me know if you have any questions!
