### 2020-08-04 #2449 @galessiorob
Hey @LukasAdamowicz wanted to check in, do you have any questions, is there anything I can do to help get your paper across the finish line?
