### 2020-08-05 #2449 @LukasAdamowicz
Only some clarification for now - for the third point - is that documentation in the paper, or as part of the package documentation?
