### 2020-08-07 #2449 @galessiorob
@gweissman I noticed that you haven't checked off the `State of the field` item from your list, the paper states that "_to the author's knowledge there are no commonly used open-source, and easy to use and configure, implementations of sit-to-stand algorithms available_", do you think this needs to be expanded?
