### 2020-08-07 #2449 @gweissman
> @gweissman I noticed that you haven't checked off the `State of the field` item from your list, the paper states that "_to the author's knowledge there are no commonly used open-source, and easy to use and configure, implementations of sit-to-stand algorithms available_", do you think this needs to be expanded?

I will check it off. I think that is sufficient although I am not knowledgeable in this field to know how true that statement is. A quick web search seems to confirm that while numerous algorithms have been published, the software does not seem to be readily available.
