### 2020-08-11 #2449 @LukasAdamowicz
Hi @galessiorob,
After a discussion with my co-author, we decided instead to remove the citation to the pending paper.

Let me know if there is anything else to do now, at this point.
