### 2020-08-15 #2449 @galessiorob
@LukasAdamowicz I think leaving the pending paper citation out is fine, at this point the next thing we need from you is to deposit your software in Zenodo and post the DOI here. After that, we can set the final version, generate a final proof and have an Editor in Chief do the publishing.
