### 2020-08-19 #2449 @kyleniemeyer
Hello @LukasAdamowicz, everything looks good, except for one minor set of changes needed in the paper: can you please add commas after "e.g." throughout? Thanks!
